finalização da autenticação e mensagens com o Exceptions

+++ src/app/Repositories/Eloquent/UserRepository.php
@@ -4,6 +4,7 @@ namespace App\Repositories\Eloquent;
 
 use App\User;
 use App\Repositories\UserRepositoryInterface;
+use Illuminate\Http\Exceptions\HttpResponseException;
 
 class UserRepository implements UserRepositoryInterface
 {
@@ -14,19 +15,28 @@ class UserRepository implements UserRepositoryInterface
 
     public function all()
     {
-        $users = $this->model->orderBy('id', 'desc')->paginate(7);
+        $users = $this->model->orderBy('id', 'desc')->paginate(9);
         return $users;
     }
 
     public function show($name)
     {
-        $users = $this->model->where('name', 'like', '%'. $name .'%')->get();
+        $users = $this->model->where('name', 'like', $name .'%')->get();
         return $users;
     }
 
     public function destroy($id)
     {
-        $this->model->destroy($id);
-        return response('Usuário excluído', 200);
+        if ($id == 1){
+            $message = [
+                'message' => 'Usuário não pode ser excluído.'
+            ];
+        } else {
+            $this->model->destroy($id);
+            $message = [
+                'message' => 'Usuário excluído.'
+            ];
+        }
+        throw new HttpResponseException(response()->json($message, 201)); 
     }
 }
